fix: keep prep wiring within file-size guard (PAN-1897)
